Realized that the standard abstract syntax tree will not be as easy to use with postfix notation. That is, instead of `1 + 2`, in my language it's `1 2 +`. I found especially difficult with the standard method, to implement a solution that can accomodate unary operators like `-1` as `-` and `1` both counts as a token.
This issue becomes even more problematic considering I cannot confirm
from just the first number (or token) if an operation is actually gonna
occur because I cannot peek two tokens ahead to make sure without making
major adjustments to the parser class.

Thus in the context of evaluating expressions, the use of the stack is
important, more so than I initially realized for postfix operations, and
so I have adjusted my expression function to accomodate this.